+++ src/constants/shipComponentTypes.js
@@ -5,7 +5,8 @@ import toMap from 'to-map'
 // Use a flat map, makes life a little easier
 const SC_TYPES = toMap({
     HULL: 'sc:hull',
-    THRUSTER: 'sc:thruster'
+    THRUSTER: 'sc:thruster',
+    TURRET: 'sc:turret'
 })